Refactor semgrep SARIFs into semgrep module
